feat: give up on pydantic-style annotation type. using a private attr instead
